Split out time ago store from component
- this way we can use it in more places
